+++ packages/Core/CHANGELOG.md
@@ -4,6 +4,11 @@
 - Use [unfetch](https://github.com/developit/unfetch), [sockette](https://github.com/lukeed/sockette) and [Websocket Node](https://github.com/theturtle32/WebSocket-Node) for new client/server code.
 - Add SDL schema support. [Link](https://www.prisma.io/blog/graphql-sdl-schema-definition-language-6755bcb9ce51/).
 
+## [2.0.5] 1-17-2018
+- Fixed issue with `nan` function.
+- Added new adapters to readme.
+- Added option to RocksDB adapter to disable caching indexes.
+
 ## [2.0.4] 1-16-2018
 - Adjusted base SQLite query function to support streams.